+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.8.0 2024-04-23
+
 ### Added
 
 - The `BulkUpload` entity now has a `createdBy` attribute.